requirements.txt
Pillow
opencv-python
numpy
requests
